requirements.txt
streamlit
pandas
requests
pandera
google-cloud-bigquery